pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "modelscope-manager"
version = "0.1.0"
description = "ModelScope Manager Project"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = ["openai", "requests", "python-dotenv"]

[tool.setuptools.packages.find]
where = ["."]
